Final simple class-based unit converter
Python-3-ic, exposes properties, subclasses dict so you can instantiate
it with some of the dictionary keys set (note that no checking of
__dict__ key setting occurs, so if you set keys that are managed by
properties that have side effects, those side effects will simply not
happen.  The assumption is that if you know enough to set state on
instantiation, you know enough to assume that any side effects that
cascade to other properties or keys will not occur).