Oops, s/itertools\,\s//
Turns out we don't need itertools, fun though it may have been to
attempt to use some of Python's iterable interface porcelain to do
something clever-looking.

What we *might* need in the future (if we're serious about Doing This
Right) may be the collections module, provided we can sanely subclass
collections.Iterator in a way that makes more sense than just binding
(in def __init__(self, *args):) args[0].__iter__ to self.__iter__, and
overloading __next__ to do some "magic" (walk through an entire search
until the substring matches >= once before returning a value, or simply
hold on to and expose more state than an iterator is supposed to in
order to give the iterator-differently-abled expression/closure making
use of it a helping hand).