Let's stop it with the hardcoded values
The length of search_s as a fixed constant still shows up twice in the
previous version of this program.  Fine if the program is *always* going
to be matching against a series of length 3, but if that changes,
breakage (especially in the case of a series of length >=4) may occour.
Plus the input validation wouldn't let the user enter in a 2-integer
series for a series_s of length 2, which would be sad.