Initial checkin of homework #7
First initial skeleton of homework #7.  Relies primarily on sets and
testing for the presence of individual members in a str() by use of
str.index(member, last_match_ind).  Defaults to reading 4096 bytes at a
time, this can be specified by the user as the second commandline argument
(after the name of a file to read).